+++ dimcli/functions.py
@@ -1,0 +1,9 @@
+"""
+Utility module to facilitate import of .core.functions
+
+``from dimcli.functions import extract_concepts``
+"""
+
+from . import *
+from .core.functions import *
+