+++ dimcli/shortcuts.py
@@ -11,6 +11,3 @@ from . import *
 
 from .utils.utils_general import *
 from .utils.utils_dimensions import *
-
-from .core.functions import *
-